Incident 488467 / 2020 Fiori Elements app stucked when inserting item lines
By default when you create a table in the ObjectPage, it is by default configure with createmode = NewPage.

As such when you click on create the applicaton tries to navigate to a dedicated "Products(...)" page

You have two options. :
- Add a dedicated product page and configure the navigation accordingly so that when you click on create you get navigated to the right page
- Change the tableSettings for this table so that it is defined as creationMode = "Inline"

To do the latter, in the manifest for the Systems objectPage add the following

"settings": {
"controlConfiguration": {
"products/@com.sap.vocabularies.UI.v1.LineItem": {
         "tableSettings": {
          "creationMode": {
           "name": "Inline"
          }
         }
        },
}
}